requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
botocore==1.29.27
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
frozenlist==1.3.3
gql==3.4.0
graphql-core==3.2.3
idna==3.4
jmespath==1.0.1
multidict==6.0.3
psycopg2-binary==2.9.5
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.13
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
